Introduce a helper for asserting events to reduce code duplication #7023
* Separate events
